requirements.txt
requests
urllib3==1.26.6
python-dateutil==2.8.2
certifi==2021.5.30
python-libnmap==0.7.2
six==1.16.0
